Add a /docs section for feature manuals
The OBS manual only existed inside a modal, so it could not be linked to,
bookmarked or browsed. Give it a home: /docs lists manuals, /docs/:manual/:section
renders one chapter at a time with its own URL.

Chapters now live in src/docs/manuals/streaming.tsx and are rendered from one
array by both the page and the existing help modal, so the two cannot drift.
The modal keeps working exactly as before and gains an 'Open full manual' link
that carries the reader to the chapter they were already on.

Named /docs rather than /guides because the blog already owns that word
(blog.woogles.io/guides, linked from the footer). The split is deliberate and
recorded in registry.ts: manuals make falsifiable claims about how the code
behaves, so they belong in the same repo and the same PR as the behaviour they
describe -- this manual has already had to be corrected twice for exactly that
reason. Editorial content has no such coupling and stays on the blog.

Adding a second manual is one entry in MANUALS.
