+++ liwords-ui/src/broadcasts/obs_help.scss
@@ -32,9 +32,6 @@
   }
 }
 
-// The OBS manual. It is a long read, so the priorities here are a comfortable
-// measure, scannable tables, and code that stands out from prose without
-// shouting.
 .obs-help-tabs {
   // Keep the chapter list visible while a long chapter scrolls under it.
   max-height: 62vh;
@@ -48,115 +45,9 @@
   }
 }
 
-.obs-help-body {
-  padding-right: 12px;
-  max-width: 62em;
-
-  @include colorModed() {
-    color: m($gray-extreme);
-  }
-}
-
-.obs-help-h {
-  margin-top: 20px !important;
-  margin-bottom: 6px !important;
-
-  &:first-child {
-    margin-top: 0 !important;
-  }
-
-  @include colorModed() {
-    color: m($gray-extreme);
-  }
-}
-
-.obs-help-p {
-  margin-bottom: 10px !important;
-  line-height: 1.6;
-
-  @include colorModed() {
-    color: m($gray-extreme);
-  }
-}
-
-.obs-help-ol,
-.obs-help-ul {
-  margin: 0 0 12px 0;
-  padding-left: 22px;
-  line-height: 1.6;
-
-  li {
-    margin-bottom: 6px;
-  }
-
-  @include colorModed() {
-    color: m($gray-extreme);
-  }
-}
-
-.obs-help-code {
-  padding: 1px 5px;
-  border-radius: 3px;
-  font-family: "Courier New", monospace;
-  font-size: 0.92em;
-  word-break: break-word;
-
-  @include colorModed() {
-    background: m($off-background);
-    color: m($gray-extreme);
-  }
-}
-
-// The pointer-chain diagram. Monospace and indentation carry the meaning, so
-// it must not wrap — it scrolls instead.
-.obs-help-diagram {
-  margin: 0 0 12px 0;
-  padding: 12px;
-  border-radius: 4px;
-  overflow-x: auto;
-  font-family: "Courier New", monospace;
-  font-size: 12px;
-  line-height: 1.5;
-  white-space: pre;
-
-  @include colorModed() {
-    background: m($off-background);
-    color: m($gray-extreme);
-  }
-}
-
-.obs-help-table {
-  width: 100%;
-  margin: 0 0 14px 0;
-  border-collapse: collapse;
-  font-size: 13px;
-  line-height: 1.5;
-
-  th,
-  td {
-    padding: 6px 10px;
-    text-align: left;
-    vertical-align: top;
-
-    @include colorModed() {
-      border-bottom: 1px solid m($gray-subtle);
-    }
-  }
-
-  th {
-    font-weight: 600;
-  }
-
-  // Leading label column in the comparison tables.
-  td:first-child {
-    white-space: nowrap;
-
-    @include colorModed() {
-      color: m($gray-medium);
-    }
-  }
-
-  @include colorModed() {
-    color: m($gray-extreme);
-  }
+.obs-help-footer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
 }
